修复bug

+++ src/main/java/com/springboot/springbootlogindemo/domain/User.java
@@ -9,7 +9,7 @@ public class User {
     // 主键自增int(10)对应long
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long uid;
+    private int uid;
 
     // 用户名属性varchar对应String
     private String uname;
@@ -17,11 +17,11 @@ public class User {
     // 密码属性varchar对应String
     private String password;
 
-    public long getUid() {
+    public int getUid() {
         return uid;
     }
 
-    public void setUid(long uid) {
+    public void setUid(int uid) {
         this.uid = uid;
     }
